scripts/requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gspread==6.2.1
html5lib==1.1
httplib2==0.22.0
idna==3.10
mechanize==0.4.10
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv>=1.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.3.0
webencodings==0.5.1
youtube-transcript-api==1.1.0
google-genai>=0.3.0
